Added InterventionVC extension which contains functions used to load lexicon. Now loading lexicon before farm's query
